specified special characters - refactor state from signup form next
